Refactor routes into blueprint
